+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 - Set WriteGroup for `Translation`, `Rotation` and `Scale`
 - Replace `TweenState.LOOP_COUNT_INFINITE` with `Tween.Infinite`
 
+### Fixed
+
+- Parallel writing with `[NativeDisableContainerSafetyRestriction]` in `TweenDestroySystem`
+
 ## [0.3.0] - 2020.11.03
 
 ### Added